api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Flower Hackathon API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.121.2",
    "uvicorn[standard]>=0.38.0",
    "sqlmodel>=0.0.22",
    "psycopg2-binary>=2.9.9",
    "python-multipart>=0.0.9",
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "torchxrayvision>=1.2.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
]
